+++ flutter_application_1/lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'routes.dart';
-import 'loginScreen.dart';
+import 'pages/loginScreen.dart';
 
 import 'package:firebase_core/firebase_core.dart';
-import 'firebase_options.dart';
+import 'services/firebase_options.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 void main() async {
