+++ flutter_application_1/lib/pages/detailswidget.dart
